Backend

See merge request s11-ai-image-sub1/S11P21B102!48

Former-commit-id: 6fd3c5214234bb8317de4a2b06d6560cdffef80a

+++ kickcap-BE/src/main/java/com/ssafy/kickcap/report/dto/RealTimeReportRequestDto.java
@@ -11,7 +11,7 @@ import java.time.ZonedDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class RealTimeReportRequestDto {
 
-    private int violationType;
+    private Long violationType;
 
     private String image;
